cli moved to separated class
