+++ core/src/main/java/pl/kbtest/UncertainRuleEngine.java
@@ -6,20 +6,20 @@
 
 package pl.kbtest;
 
-import pl.kbtest.contract.Context;
-
-import java.math.BigDecimal;
-import java.util.Deque;
-import java.util.LinkedList;
-import java.util.List;
-
 import pl.kbtest.conclusionExecutor.SetConclusionExecutor1;
+import pl.kbtest.contract.Context;
 import pl.kbtest.contract.GrfIrf;
 import pl.kbtest.contract.SetFact;
 import pl.kbtest.contract.SetFactFactory;
 import pl.kbtest.contract.SetPremise;
 import pl.kbtest.contract.SetRule;
 import pl.kbtest.premiseEvaluator.SetPremiseComparator;
+import pl.kbtest.utils.Utils;
+
+import java.math.BigDecimal;
+import java.util.Deque;
+import java.util.LinkedList;
+import java.util.List;
 
 /**
  * @author Kamil
